refactor(actions): split execute_action into domain sub-dispatchers
execute_action was a single ~900-line match with 40+ arms mixing
terminal, tab, git, files and project ops. ActionRequest is a flat enum,
so the top-level match stays exhaustive (no unreachable! catch-alls) but
each arm now destructures and delegates to a grouped handler in a new
execute/ submodule: terminal.rs, tab.rs, git.rs, files.rs, project.rs.
execute_action becomes a thin routing table.

Pure reorganization (git mv preserves history): signature unchanged,
671 tests pass identically, no behavior or API change.
